requirements.txt
python-telegram-bot==20.7
httpx>=0.25.0
aiosqlite>=0.19.0
asyncssh>=2.14.0
cryptography>=41.0.0
